Refactor datagroup service and UI
