+++ .gitignore
@@ -14,3 +14,9 @@
 .cxx
 local.properties
 app/src/main/res/xml/network_security_config.xml
+app/src/main/kotlin/de/schnitzel/shelfify/api/ApiConfig.kt
+.idea/copilot.data.migration.agent.xml
+.idea/copilot.data.migration.ask.xml
+.idea/copilot.data.migration.edit.xml
+.idea/copilot.data.migration.ask2agent.xml
+app/src/main/kotlin/de/schnitzel/shelfify/api/ApiConfig.kt
