+++ app/src/main/kotlin/de/schnitzel/shelfify/funktionen/sub/DatagroupService.kt
@@ -102,7 +102,7 @@ object DatagroupService {
 
                 withContext(Dispatchers.Main) {
                     when (code) {
-                        200 -> Toast.makeText(activity, "Einladung erfolgreich!", Toast.LENGTH_SHORT).show()
+                        200 -> Toast.makeText(activity, "Verlassen erfolgreich!", Toast.LENGTH_SHORT).show()
                         400 -> Toast.makeText(activity, "Ungültige Anfrage!", Toast.LENGTH_SHORT).show()
                         401 -> Toast.makeText(activity, "Token ungültig!", Toast.LENGTH_SHORT).show()
                         else -> Toast.makeText(activity, "Fehler: $code", Toast.LENGTH_SHORT).show()
